Extract restart action
